測定データの表示用コンポーネントを整備
